refactor(sap-aicore-embedder): extract decodeEmbedding helper for reuse
